requirements.txt
azure-core==1.28.0
azure-identity==1.13.0
certifi==2023.5.7
cffi==1.15.1
cryptography==41.0.2
idna==3.4
microsoft-kiota-abstractions==0.5.1
microsoft-kiota-authentication-azure==0.2.0
microsoft-kiota-http==0.4.3
microsoft-kiota-serialization-json==0.3.4
microsoft-kiota-serialization-text==0.2.0
msal==1.22.0
msal-extensions==1.0.0
msgraph-core==1.0.0a4
msgraph-sdk==1.0.0a12
pycparser==2.21
PyJWT==2.7.0
python-dateutil==2.8.2
requests==2.31.0
six==1.16.0
uritemplate==4.1.1
urllib3==2.0.3
